Fixed tests

+++ src/main/java/it/polimi/ingsw/PSP18/server/controller/MatchRun.java
@@ -136,8 +136,10 @@ public class MatchRun {
             }
 
             // Cancel the backup file of the match because the match has ended
-            File f = new File(match.getBackupManager().getFileName());
-            f.delete();
+            if(match.getBackupManager().getFileName() != null) {
+                File f = new File(match.getBackupManager().getFileName());
+                f.delete();
+            }
         } else {
             for(SocketThread sock : match.getMatchSocket().getSockets()) {
                 sock.closeConnection();
